adding separate blobs view
