+++ django_oci/settings.py
@@ -60,7 +60,9 @@ STORAGE_BACKEND = oci.get("STORAGE_BACKEND", DEFAULTS["STORAGE_BACKEND"])
 DOMAIN_URL = oci.get("DOMAIN_URL", DEFAULTS["DOMAIN_URL"])
 MEDIA_ROOT = oci.get("MEDIA_ROOT", DEFAULTS["MEDIA_ROOT"])
 CACHE_DIR = oci.get("CACHE_DIR", DEFAULTS["CACHE_DIR"])
-DISABLE_TAG_MANIFEST_DELETE = oci.get("DISABLE_TAG_MANIFEST_DELETE", DEFAULTS["DISABLE_TAG_MANIFEST_DELETE"])
+DISABLE_TAG_MANIFEST_DELETE = oci.get(
+    "DISABLE_TAG_MANIFEST_DELETE", DEFAULTS["DISABLE_TAG_MANIFEST_DELETE"]
+)
 
 SESSION_EXPIRES_SECONDS = oci.get(
     "SESSION_EXPIRES_SECONDS", DEFAULTS["SESSION_EXPIRES_SECONDS"]
