+++ django_oci/signals.py
@@ -19,6 +19,7 @@ from django.db.models.signals import post_delete
 from django.dispatch import receiver
 from .models import Image
 
+
 @receiver(post_delete, sender=Image)
 def delete_blobs(sender, instance, **kwargs):
     print("Delete image signal running.")
