+++ django_oci/views/__init__.py
@@ -1,3 +1,9 @@
 from .base import APIVersionCheck
-from .image import ImageManifest, ImageBlobUpload, ImageBlobDownload
+from .blobs import BlobUpload, BlobDownload
+from .image import ImageManifest, ImageTags
 from .chunked import ChunkedUploadDemo, MyChunkedUploadView, MyChunkedUploadCompleteView
+
+# Load storage on application init
+from django_oci.storage import get_storage
+
+storage = get_storage()
